Moved PluginViewModel to its own file
